Updated notebook with visualization and comparison of models.
